requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic>=0.25.0
python-dotenv==1.0.0
